+++ Lab4/Component.hpp
@@ -6,7 +6,7 @@
 
 struct Connection {
     //contain current charge of connection
-    float charge{};
+    double charge{};
 };
 
 class Component {
@@ -14,9 +14,9 @@ public:
     Component(std::string name, Connection& a, Connection& b);
     virtual ~Component() = default;
     virtual void simulate() const = 0;
-    virtual void set_connection_values(float& timestep) = 0;
-    virtual float get_current() const = 0;
-    float get_voltage() const;
+    virtual void set_connection_values(double& timestep) = 0;
+    virtual double get_current() const = 0;
+    double get_voltage() const;
     virtual std::string get_name() const = 0;
 protected:
     //every component has a and b and name
