+++ Lab4/Resistor.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-Resistor::Resistor(string name, float resistance, Connection& a, Connection& b) : Component(name, a, b), resistance{resistance} {}
+Resistor::Resistor(string name, double resistance, Connection& a, Connection& b) : Component(name, a, b), resistance{resistance} {}
 
 void Resistor::simulate() const {
     cout << "resistorar SIM" << endl;
@@ -12,15 +12,15 @@ string Resistor::get_name() const {
     return name;
 }
 
-float Resistor::get_current() const {
+double Resistor::get_current() const {
         //voltage / resistance
-    float voltage{get_voltage()};
+    double voltage{get_voltage()};
     //cout << "resistor current" << voltage/resistance <<endl;
     return voltage/resistance;
 }
 
 //resistor will move charge from most charged to least charged
-void Resistor::set_connection_values(float& timestep) {
+void Resistor::set_connection_values(double& timestep) {
     //a is 5, b is 9
     if(a.charge < b.charge){
         a.charge += (abs(a.charge - b.charge)/resistance) * timestep;
